requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchaudio>=2.0.0
torchcodec>=0.10.0
pyannote.audio>=3.0.0
PyQt5
